Use leave instruction. Fix builds.
Inspection of GCC output reveals the use of the 'leave' instruction
in place of the following:

	mov esp, ebp
	pop ebp

This should probably be favoured in CoreLoader as well to reduce
amount of explicit instructions and receive internal CPU
optimisations found in the instruction.

However the counterpart 'enter' instruction is _not_ used by GCC.
Research on this reveals that the 'enter' instruction has some
major performance issues.

+++ libc/stdio/vsnprintf.c
@@ -299,13 +299,13 @@ static uint32_t __kd_sn_to_str(char *restrict str, int64_t num, uint8_t base)
 	}
 
 	uint8_t neg = (num < 0);
-	register uint64_t v = num * (neg ? -1 : 1);
-	register uint64_t dV = 0;
+	register int64_t v = num * (neg ? -1 : 1);
+	register int64_t dV = 0;
 	char buffer[64] = { 0 };
 	register char *ptr = buffer + 62;
 	register uint32_t len = 1;
 
-	while (v >= (uint64_t)base) {
+	while (v >= base) {
 		dV = v % base;
 		v /= base;
 		*ptr-- = *(digits + dV);
